Improve selected block logic in Lexical
